+++ packages/ethers/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "ethers/5.0.29";
+export declare const version = "ethers/5.0.30";
 //# sourceMappingURL=_version.d.ts.map