+++ packages/providers/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "providers/5.0.21";
+export declare const version = "providers/5.0.22";
 //# sourceMappingURL=_version.d.ts.map